Integrate Document Classification into identiDoc (#17)
The classification module is now fully integrated into the identiDoc system. identiDoc no longer triggers a download of a txt file of extracted text, now the user is shown a browser alert with the appropriate classification. txt files are no longer accepted by identiDoc - a txt file is not an appropriate format for a document.

Co-authored-by: Sandesh Koirala <sandesh.koirala@mavs.uta.edu>